get cruner app info after downloading app

+++ src/cli/helpers/remove_app.ts
@@ -1,10 +1,15 @@
 import { Config } from "../config";
 import { join } from "path";
-import { unlink, remove } from "fs-extra";
+import { unlink, remove, pathExists } from "fs-extra";
+import { Logger } from "../../commons";
 export async function removeApp(appName: string) {
     const installDir = Config.installDir;
     const appinstallDir = join(installDir, appName);
     const appBinDir = join(Config.binDir, appName);
-    await unlink(appBinDir);
+    const isAppExist = await pathExists(appBinDir);
+    if (isAppExist) {
+        await unlink(appBinDir);
+    }
     await remove(appinstallDir);
+    Logger.debug(`Installed app ${appName} removed`)
 }
